chore: improve Clippy on CI
Clippy is run on CI, but even if there are warnings, the CI doesn't
fail. With this commit the CI fails when there are Clippy warnings.
It also runs Clippy on all the code (also tests, benchmarks and
examples) now.

This commit also removes the build targets as they are no covered by
the Clippy runs, there is no need to build things again.