fix: prevent potential deadlock (#200)
Spawn the thread that is receiving messages outside the Rayon thread
pool, so that it can always receive messages, even if the thread pool
is fully occupied. If many senders are spawn before the receiver, there
will be a deadlock.

To illustrate the issue, run the `test_groth16_aggregation` with a
thread pool of size 1. Prior to this change, the test deadlocks:

    RAYON_NUM_THREADS=1 cargo test --release --tests test_groth16_aggregation -- --exact --nocapture