fix: make it work with latest GPU kernels (#278)
The FFT and Multiexp kernels operates on fields instead of an engine.
For Multiexp it means that for BLS12-381 we need two separate kernels,
one for G1 and one for G2.

The kernel generation is no longer set via feature flags in `ec-gpu-gen`,
but instead it is directly specified in this library in the `build.rs`.

BREAKING CHANGE: upgrade to blstrs 0.6.0

`blstrs` version 0.6.0 is pinned to `blst` 0.3.10. All dependencies
need to be at the same version. Hence doing a breaking change release
is needed.