fix: remove wasm and related features (#281)
The `wasm` feature was introduced as `getrandom` needs the `js` feature
in order to compile for `wasm32-unknown-unknown`. A better fix is to disable
the default features of `rand` instead. This way the `wasm` feature is no
longer needed.

For compiling on WebAssembly also the `memmap` feature was introduced. Remove
that feature and replace it with target architecture flags, that exclude
memmap related functions when compiled for wasm.

This lowers the maintenance burden as WebAssembly is no longer a special case.

Thanks @samuelburnham for the code that made the removal of `memmap` feature
possible.

BREAKING CHANGE: `wasm` and `memmap` features are removed

If you're targetting WASM and using `bellperson` as a dependency, prior to this
change you'd have included it as:

    bellperson = { version = "0.23", default-features = false, features = ["wasm"] }

Now you can instead just rely on the default features:

    bellperson = "0.23"

BREAKING CHANGE: resolver v2 is used

In order to make things compile nicely for the wasm target, the resolver was changed
to version 2.