chore: remove test_utils module
`test_utils` is no longer used, due to parts of the code being moved
into ec-gpu project instead.

BREAKING CHANGE: it was part of the public test interface

Technically it's a breaking change, as the module was part of the public
test interface. Practically this shouldn't have an impact on anyone.