fix: make priority locking work again (#293)
Due to refactorings, the `PriorityLock::should_break()` logic was
quite confusing and used the wrong way. Make it work correctly
while simplifying the logic.

This commit also removes `PriorityLock` from the public API as it
isn't really needed.

Tweak some values in the parallel prover test, so that aborting a
low priority kernel from running on the GPU happens more frequently.

It needs a newer version of `ec-gpu-gen`, else it would cause panics
(which are not fatal, as they happen within a thread. Though, they
still show up in the logs).

Fixes #291.