chore: increase MSRV to 1.62.1 (#302)
Set the minimum supported Rust version to 1.61.1. Technically it's not
needed as the bump is only needed for the `temp-env` dev dependency.
Nonetheless is 1.62.1 not that new anymore, hence I'm bumping the
version so that we don't need to special case the version for testing.

Also fix all Clippy warnings that were introduced in the new Rust version.