feat: add WitnessCS, SizedWitness, and related ConstraintSystem methods. (#309)
This PR adds some new ConstraintSystem methods. These methods are intended to
only be implemented by or called on ConstraintSystems that are also designated
as witness generators. Such ConstraintSystems will signal this by implementing
and returning true from `is_witness_generator()`.

A minimal `WitnessCS` suitable for this purpose is also provided, as is a
`SizedWitness` trait which should be implemented by objects (for example,
implementers of `Circuit`) that choose to offer efficient witness-generation.

In this context, 'efficient witness generation' likely bypasses the usual,
flexible mechanism used to specify `bellperson` circuits. Although convenient,
this mechanism conflates synthesis and witness generation in ways that can add
significant overhead.

Initial synthesis benchmarks in Neptune, although imperfect, showed 15x speedup
when using this method, and this PR's immediate purpose is to support that use
case.